Correctly quote post/draft/page names
This fixes an inconsistency with how values are quoted for the Octopress
shell command, allowing for posts, drafts, and pages to include double
quotes in their names.

As a sort-of bonus, names not containing double quotes are quoted for
the shell with the standard `shell-quote-argument`, which is probably
more robust for all other cases.