Add interactive function for inserting image path
This is helpful for inserting image tags so that you don't have to
type out a whole relative path, which Emacs generally won't know how
to autocomplete (because it is relative).

This function simply subtracts the root of the Octopress project from
the beginning of the path. It should fit most use cases.